% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{append_db}
\alias{append_db}
\title{Append sf geometry table to a database table}
\usage{
append_db(con, x, tableName)
}
\arguments{
\item{con}{database connection object from connect_pg}

\item{x}{sf object to write}

\item{tableName}{name of table; can use DBI::SQL use non-default schema}
}
\description{
Append sf geometry table to a database table
}
