% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{bcgov_theme}
\alias{bcgov_theme}
\title{BCGov Shiny theme}
\format{
An object of class \code{bs_bootswatch_materia} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
bcgov_theme
}
\description{
A function to create a theme for shiny
}
\keyword{datasets}
