% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{BCGovColors}
\alias{BCGovColors}
\alias{bcgov_primary}
\alias{bcgov_text}
\alias{bcgov_links}
\alias{bcgov_components}
\alias{bcgov_semantic_colors}
\title{BCGov Developer color palettes}
\usage{
bcgov_primary()

bcgov_text()

bcgov_links()

bcgov_components()

bcgov_semantic_colors()
}
\value{
vector of hex codes
}
\description{
See \url{https://developer.gov.bc.ca/Design-System/Colour-Palette}
}
