% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{card}
\alias{card}
\title{Wrapper to create bootswatch card}
\usage{
card(..., title, class = "", style = "", headerClass = "bg-primary")
}
\arguments{
\item{...}{arguments to pass to card body div}

\item{title}{title of card header}

\item{class}{additional classes to add}

\item{style}{additional styles to add to card}

\item{headerClass}{additional classes to add to header only}
}
\description{
Wrapper to create bootswatch card
}
