% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{check_req_shp_files}
\alias{check_req_shp_files}
\title{Check if required files for shapefiles are present}
\usage{
check_req_shp_files(selectedFile, uploadedFiles)
}
\arguments{
\item{selectedFile}{selected file with .shp file extension}

\item{uploadedFiles}{vector of all uploaded files}
}
\value{
boolean
}
\description{
Check if required files for shapefiles are present
}
