% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_pg}
\alias{connect_pg}
\title{Create connection to postgres for shiny app}
\usage{
connect_pg(
  dbname,
  host = Sys.getenv("PGHOST"),
  port = 5432,
  user = Sys.getenv("PGUSER"),
  password = Sys.getenv("PGPASSWORD")
)
}
\arguments{
\item{dbname}{database name}

\item{host}{server/host of postgres database}

\item{port}{port number}

\item{user}{user name}

\item{password}{password}
}
\value{
DBI::dbConnect object for local development and pool::dbPool object for
deployment
}
\description{
Create connection to postgres for shiny app
}
