% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{convert_to_image}
\alias{convert_to_image}
\title{Converts binary data of photo in database to an on-disk image}
\usage{
convert_to_image(con, id)
}
\arguments{
\item{con}{connection to postgres}

\item{id}{unique data point id}
}
\value{
name of temp file where image is stored
}
\description{
Converts binary data of photo in database to an on-disk image
}
