% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbOperations}
\alias{dbOperations}
\alias{transfer_field_data_points}
\title{PEM database operations}
\usage{
transfer_field_data_points(con, transectIds)
}
\arguments{
\item{con}{connection to PEM database}

\item{transectIds}{transect ids}
}
\description{
PEM database operations
}
