% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbPhotoUi}
\alias{dbPhotoUi}
\alias{dbPhoto}
\alias{dbPhotoServer}
\title{Module to select a point and upload a photo for it}
\usage{
dbPhotoUi(id)

dbPhotoServer(id, con)
}
\arguments{
\item{id}{namespace id}

\item{con}{connection to postgres database}
}
\value{
a shiny reactive object row id of photo and a database trigger
}
\description{
Module to select a point and upload a photo for it
}
