% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbWrite}
\alias{dbWrite}
\alias{dbWriteUi}
\alias{dbWriteServer}
\title{Database upload module}
\usage{
dbWriteUi(id)

dbWriteServer(id, sfObject, tableName, con)
}
\arguments{
\item{id}{namespace id}

\item{sfObject}{shiny reactive sf object}

\item{tableName}{name of table; can use DBI::SQL use non-default schema}

\item{con}{database connection object from connect_pg}
}
\description{
Database upload module
}
