% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{find_boundary_point}
\alias{find_boundary_point}
\title{Find boundary point of geometry}
\usage{
find_boundary_point(sfObject, coord = c("lat", "lng"), type = c("min", "max"))
}
\arguments{
\item{sfObject}{sf data.frame}

\item{coord}{which axis to find extreme point}

\item{type}{which direction min/max}
}
\value{
sf dataframe
}
\description{
Find boundary point of geometry
}
