% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geomUpload}
\alias{geomUpload}
\alias{geomUploadUi}
\alias{geomUploadServer}
\title{Shiny Module to upload geometry}
\usage{
geomUploadUi(id)

geomUploadServer(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
reactive value of sf object
}
\description{
Shiny Module to upload geometry
}
