% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_point_id}
\alias{get_point_id}
\title{Get the unique data point id from staging/main table}
\usage{
get_point_id(con, transectId, observer, order)
}
\arguments{
\item{con}{connection to postgres}

\item{transectId}{transect id}

\item{observer}{observer}

\item{order}{point order of sampling}
}
\value{
integer
}
\description{
Get the unique data point id from staging/main table
}
