% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{guess_geometry_type}
\alias{guess_geometry_type}
\title{Guess the geometry type of  a sf or sfc object}
\usage{
guess_geometry_type(sfObject)
}
\arguments{
\item{sfObject}{sf dataframe object}
}
\value{
character vector size 1 with geometry type
}
\description{
Guess the geometry type of  a sf or sfc object
}
