% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{is_in_db}
\alias{is_in_db}
\title{Check if transect is already in the database}
\usage{
is_in_db(con, transectIds, observers, stagingTable, transectsTable)
}
\arguments{
\item{con}{connection to postgres}

\item{transectIds}{vector of transect ids}

\item{observers}{vector of observers}

\item{stagingTable}{name of staging table \link{staging_tables}}

\item{transectsTable}{name of transects table \link{transects_tables}}
}
\value{
boolean vector
}
\description{
Check if transect is already in the database
}
