% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{make_reactive_trigger}
\alias{make_reactive_trigger}
\title{Make a reactive trigger}
\usage{
make_reactive_trigger()
}
\description{
A function to create a trigger to place inside reactive elements.
}
\examples{
triggerOne <- make_reactive_trigger()
# place triggerOne$depends() in reactive element to update
# place triggerOne$trigger() in code that should force updates
}
