% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PEMData}
\alias{PEMData}
\alias{map_unit}
\alias{point_type}
\alias{transition}
\alias{struc_mod}
\alias{data_type}
\alias{struc_change}
\title{Functions to produce a list of categories for PEM}
\usage{
map_unit()

point_type()

transition()

struc_mod()

data_type()

struc_change()
}
\value{
named list with code as name and description as value
}
\description{
Functions to produce a list of categories for PEM
}
