% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{PEMmapUi}
\alias{PEMmapUi}
\alias{mapUi}
\alias{mapServer}
\title{Shiny module to show PEM data}
\usage{
mapUi(id)

mapServer(id, con, sfObject, success)
}
\arguments{
\item{id}{namespace id}

\item{con}{connection to postgres database}

\item{sfObject}{uploaded sf object}

\item{success}{shiny db trigger}
}
\description{
Shiny module to show PEM data
}
