% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_PEM_app}
\alias{run_PEM_app}
\title{Run shiny app in package}
\usage{
run_PEM_app(local = TRUE)
}
\arguments{
\item{local}{TRUE if developing locally. FALSE uses the app from the installed package
location.}
}
\description{
Run shiny app in package
}
