% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runPEMApp}
\alias{runPEMApp}
\title{Run shiny app in package}
\usage{
runPEMApp(local = TRUE)
}
\arguments{
\item{local}{TRUE if developing locally. FALSE uses the app from the installed package
location.}
}
\description{
Run shiny app in package
}
