% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ShowPhoto}
\alias{ShowPhoto}
\alias{dbShowPhotoUi}
\alias{dbShowPhotoServer}
\title{Queries a photo from the database and shows the image}
\usage{
dbShowPhotoUi(id)

dbShowPhotoServer(id, con, selectedPoint)
}
\arguments{
\item{id}{namespace id}

\item{con}{connection to postgres}

\item{selectedPoint}{reactive list with id of point in database and a database trigger}
}
\description{
Queries a photo from the database and shows the image
}
