% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{switchInput}
\alias{switchInput}
\alias{updateSwitchInput}
\title{Switch Input Control}
\usage{
switchInput(
  inputId,
  label,
  value = FALSE,
  disabled = FALSE,
  width = NULL,
  class = ""
)

updateSwitchInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (TRUE or FALSE).}

\item{disabled}{disable user input}

\item{width}{The width of the input, e.g. '400px', or '100\%'; see validateCssUnit().}

\item{class}{additional classes for form.control}

\item{session}{The session object passed to function given to shinyServer. Default is
getDefaultReactiveDomain().}
}
\value{
A checkbox control that can be added to a UI definition.
}
\description{
Switch Input Control
}
