% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{upsert_photo}
\alias{upsert_photo}
\title{Add/Update photo to photo table}
\usage{
upsert_photo(con, photoDf)
}
\arguments{
\item{con}{connection to postgres}

\item{photoDf}{dataframe with photo as a blob}
}
\value{
error or TRUE if successful
}
\description{
Add/Update photo to photo table
}
