% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate_data}
\alias{validate_membership}
\alias{validate_character}
\alias{validate_float}
\alias{validate_integer}
\alias{validate_date}
\alias{validate_time}
\alias{validate_missing}
\alias{validate_nothing}
\alias{validate_geometry}
\alias{validate_geometry_type}
\alias{validate_contains}
\alias{validate_char_length}
\alias{validate_transect_id}
\alias{validate_map_unit_change}
\alias{validate_map_unit_transition}
\alias{validate_points_column}
\alias{validate_PEM_data}
\alias{validate_field_points_data}
\alias{validate_tracklog_column}
\alias{validate_field_tracklog_data}
\alias{err_column_name}
\alias{validate_length}
\alias{validate_sample_pairs}
\title{Validation tests for vectors}
\usage{
validate_data(x, ...)

validate_membership(members, optional = FALSE)

validate_character()

validate_float()

validate_integer()

validate_date()

validate_time(optional = FALSE)

validate_missing()

validate_nothing()

validate_geometry()

validate_geometry_type(geometryType)

validate_contains(string, times = 1L, exact = FALSE)

validate_char_length(maxLength)

validate_transect_id()

validate_map_unit_change(mapunit1, mapunit2)

validate_map_unit_transition(mapunit2)

validate_points_column(dataFrame, colName)

validate_PEM_data(dataFrame, f, progress, n = 1)

validate_field_points_data(transectPoints, progress)

validate_tracklog_column(dataFrame, colName)

validate_field_tracklog_data(transectLines, progress, n = 1)

err_column_name(colName)

validate_length(n, distinct = TRUE)

validate_sample_pairs(transectIds)
}
\arguments{
\item{x}{vector of values to test}

\item{...}{validation functions for x. See example for writing custom tests.}

\item{members}{vector of values to test membership}

\item{optional}{TRUE/FALSE for whether missing values are acceptable}

\item{geometryType}{geometry type to validate}

\item{string}{character to vector to match}

\item{times}{integer exact number of expected matches}

\item{exact}{TRUE matches the exact times arg and FALSE is >= times}

\item{maxLength}{integer maximum allowable character length}

\item{mapunit1}{first map unit}

\item{mapunit2}{second map unit}

\item{dataFrame}{data.frame to validate}

\item{colName}{character vector size 1 of column to validate}

\item{f}{a column validation function for PEM data}

\item{progress}{optional shiny progress object}

\item{n}{length}

\item{transectPoints}{sf object with transect field points data}

\item{transectLines}{sf object of tracklog line data}

\item{distinct}{TRUE/FALSE should test be based on unique values}

\item{transectIds}{vector of transect ids}
}
\value{
list with status of pass/fail and message of error
}
\description{
Validation tests for vectors
}
\examples{

# usage
validate_data(x = c('A', 'B'), validate_character(), validate_integer())

# custom validation function
validate_complex <- function() {
function(x) {
  success <- is.complex(x)
  description <- ifelse(success, '', 'Invalid complex type.')
  list(status = success, message = description)
}
}
validate_data(complex(rep(1, 2), c(1, -1)), validate_complex())
}
