% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validatePairs}
\alias{validatePairs}
\alias{validatePairsUi}
\alias{validatePairsServer}
\title{Shiny Module to validate transect pairs}
\usage{
validatePairsUi(id)

validatePairsServer(id, sfObject, success, con)
}
\arguments{
\item{id}{namespace id}

\item{sfObject}{reactive object returning sf data.frame}

\item{success}{reactive trigger when data is successfully uploaded}

\item{con}{connection to postgres database}
}
\description{
Shiny Module to validate transect pairs
}
