% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbOperations}
\alias{dbOperations}
\alias{transfer_field_data_points}
\alias{transfer_field_tracklog}
\title{PEM database operations}
\usage{
transfer_field_data_points(con, transectIds)

transfer_field_tracklog(con, stagedIds)
}
\arguments{
\item{con}{connection to PEM database}

\item{transectIds}{transect ids}

\item{stagedIds}{integer id of staged tracklogs}
}
\description{
PEM database operations
}
