% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate_data}
\alias{validate_membership}
\alias{validate_character}
\alias{validate_float}
\alias{validate_integer}
\alias{validate_date}
\alias{validate_time}
\alias{validate_missing}
\alias{validate_nothing}
\alias{validate_geometry}
\title{Validation tests for vectors}
\usage{
validate_data(x, ...)

validate_membership(members)

validate_character()

validate_float()

validate_integer()

validate_date()

validate_time(x)

validate_missing()

validate_nothing()

validate_geometry()
}
\arguments{
\item{x}{vector of values to test}

\item{...}{validation functions for x. See example for writing custom tests.}

\item{members}{vector of values to test membership}
}
\value{
list with status of pass/fail and message of error
}
\description{
Validation tests for vectors
}
\examples{

# usage
validate_data(x = c('A', 'B'), validate_character(), validate_integer())

# custom validation function
validate_complex <- function() {
function(x) {
  success <- is.complex(x)
  description <- ifelse(success, '', 'Invalid complex type.')
  list(status = success, message = description)
}
}
validate_data(complex(rep(1, 2), c(1, -1)), validate_complex())
}
