% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateTableUi}
\alias{validateTableUi}
\alias{validateTable}
\alias{validateTableServer}
\title{Shiny module to validate and report}
\usage{
validateTableUi(id)

validateTableServer(id, sfObject, success, con)
}
\arguments{
\item{id}{namespace id}

\item{sfObject}{reactive object returning sf dataframe}

\item{success}{reactive trigger when data is successfully uploaded}

\item{con}{connection to postgres database}
}
\description{
Shiny module to validate and report
}
