% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateTable}
\alias{validateTable}
\alias{validateTableUi}
\alias{validateTableServer}
\alias{validatePairsUi}
\alias{validatePairsServer}
\title{Shiny Module to validate and report}
\usage{
validateTableUi(id)

validateTableServer(id, sfObject)

validatePairsUi(id)

validatePairsServer(id, sfObject)
}
\arguments{
\item{id}{namespace id}

\item{sfObject}{reactive object returning sf data.frame}
}
\description{
Shiny Module to validate and report

Shiny Module to validate transect pairs
}
